Candle and helpers to dedicated file
