+++ tests/core/test_candle.py
@@ -3,26 +3,16 @@ import logging
 
 import pytest
 
-from athena.core.market_entities import Candle
-from athena.core.fluctuations import (
-    _convert_candles_to_period,
-    _merge_candles,
-    _sanitize_candles,
+from athena.core.candle import (
+    convert_candles_to_period,
+    sanitize_candles,
+    merge_candles,
     load_candles_from_file,
-    Fluctuations,
 )
-from athena.core.types import Coin, Period
+from athena.core.fluctuations import Fluctuations
+from athena.core.types import Period
 from athena.testing.equality import assert_candles_equal
 from athena.testing.generate import generate_candles
-
-
-def test_load_candles_from_file(tmp_path):
-    candles = generate_candles(size=10)
-    Fluctuations.from_candles(candles).save(tmp_path / "fluctuations.csv")
-    for candle, expected_candle in zip(
-        load_candles_from_file(tmp_path / "fluctuations.csv"), candles
-    ):
-        assert_candles_equal(candle, expected_candle)
 
 
 def test_merge_candles():
@@ -32,7 +22,7 @@ def test_merge_candles():
         period=Period(timeframe="1m"), from_date=from_date, to_date=to_date
     )
 
-    merged_candle = _merge_candles(candles)
+    merged_candle = merge_candles(candles)
 
     for candle in candles:
         assert candle.open_time >= merged_candle.open_time
@@ -52,7 +42,7 @@ def test_convert_candles_to_same_period():
     )
     target_period = Period(timeframe="1m")
 
-    merged_candles = _convert_candles_to_period(candles, target_period)
+    merged_candles = convert_candles_to_period(candles, target_period)
 
     assert len(merged_candles) == len(candles)
 
@@ -66,9 +56,13 @@ def test_convert_candles_to_period():
     )
     target_period = Period(timeframe="4h")
 
-    merged_candles = _convert_candles_to_period(candles, target_period)
+    merged_candles = convert_candles_to_period(candles, target_period)
 
     assert len(merged_candles) == 6
+
+
+def _convert_candles_to_period(candles, target_period):
+    pass
 
 
 def test_convert_candles_to_period_missing_data(caplog):
@@ -82,7 +76,7 @@ def test_convert_candles_to_period_missing_data(caplog):
 
     logging.getLogger().setLevel(logging.DEBUG)
 
-    merged_candles = _convert_candles_to_period(candles, target_period)
+    merged_candles = convert_candles_to_period(candles, target_period)
 
     assert len(merged_candles) == 6
     assert "Last candle could not be closed, won't be kept." in caplog.text
@@ -98,37 +92,7 @@ def test_convert_candles_to_period_wrong_timeframe():
     target_period = Period(timeframe="1m")
 
     with pytest.raises(ValueError, match="Cannot convert candles to lower timeframe"):
-        _convert_candles_to_period(candles, target_period)
-
-
-@pytest.fixture
-def raw_candles():
-    other_params = {
-        "open": 100,
-        "close": 100,
-        "high": 150,
-        "low": 70,
-        "coin": Coin.BTC,
-        "currency": Coin.USDT,
-        "period": "1h",
-        "open_time": datetime.datetime(2024, 1, 1),
-        "close_time": datetime.datetime(2024, 1, 2),
-        "quote_volume": 150,
-        "nb_trades": 100,
-        "taker_volume": 50,
-        "taker_quote_volume": 70,
-    }
-
-    def _raw_candles(volumes: list[int]) -> list[Candle]:
-        return [
-            Candle(
-                volume=volume,
-                **other_params,
-            )
-            for volume in volumes
-        ]
-
-    return _raw_candles
+        convert_candles_to_period(candles, target_period)
 
 
 def test_sanitize_candles():
@@ -137,13 +101,22 @@ def test_sanitize_candles():
     candles = generate_candles(size=5)
     candles[2].volume = 0
     candles[4].volume = 0
-    sanitized_candles = _sanitize_candles(candles)
+    sanitized_candles = sanitize_candles(candles)
 
     assert len(sanitized_candles) == 3
     assert all([candle.volume != 0 for candle in sanitized_candles])
 
 
 def test_sanitize_candles_empty():
-    sanitized_candles = _sanitize_candles([])
+    sanitized_candles = sanitize_candles([])
 
     assert sanitized_candles == []
+
+
+def test_load_candles_from_file(tmp_path):
+    candles = generate_candles(size=10)
+    Fluctuations.from_candles(candles).save(tmp_path / "fluctuations.csv")
+    for candle, expected_candle in zip(
+        load_candles_from_file(tmp_path / "fluctuations.csv"), candles
+    ):
+        assert_candles_equal(candle, expected_candle)
